+++ src/main/groovy/projects/patientfinder/hull/encounter.groovy
@@ -2,16 +2,28 @@ package projects.patientfinder.hull
 
 import ca.uhn.fhir.model.api.TemporalPrecisionEnum
 import de.kairos.centraxx.fhir.r4.utils.FhirUrls
+import de.kairos.fhir.centraxx.metamodel.CrfTemplateField
 import de.kairos.fhir.centraxx.metamodel.Episode
+import de.kairos.fhir.centraxx.metamodel.LaborFinding
+import de.kairos.fhir.centraxx.metamodel.LaborFindingLaborValue
+import de.kairos.fhir.centraxx.metamodel.LaborMapping
+import de.kairos.fhir.centraxx.metamodel.LaborMethod
+import de.kairos.fhir.centraxx.metamodel.LaborValue
 import de.kairos.fhir.centraxx.metamodel.Multilingual
+import de.kairos.fhir.centraxx.metamodel.OrganisationUnit
 import de.kairos.fhir.centraxx.metamodel.PatientTransfer
+import de.kairos.fhir.centraxx.metamodel.ValueReference
+import de.kairos.fhir.centraxx.metamodel.enums.EpisodeStatus
+import org.hl7.fhir.r4.model.CatalogEntry
 import org.hl7.fhir.r4.model.Encounter
 
 import static de.kairos.fhir.centraxx.metamodel.AbstractCode.CODE
+import static de.kairos.fhir.centraxx.metamodel.AbstractCodeSyncIdMultilingual.MULTILINGUALS
 import static de.kairos.fhir.centraxx.metamodel.AbstractIdContainer.ID_CONTAINER_TYPE
 import static de.kairos.fhir.centraxx.metamodel.AbstractIdContainer.PSN
 import static de.kairos.fhir.centraxx.metamodel.IdContainerType.DECISIVE
 import static de.kairos.fhir.centraxx.metamodel.RootEntities.episode
+import static de.kairos.fhir.centraxx.metamodel.RootEntities.medication
 
 /**
  * Represents a CXX Episode.
@@ -25,6 +37,20 @@ import static de.kairos.fhir.centraxx.metamodel.RootEntities.episode
  * @author Mike Wähnert
  * @since v.1.13.0, CXX.v.2023.3.0
  */
+
+final String TREATMENT_SERVICE = "treatmentService"
+final String TYPE = "type"
+final String DISCHARGE_DISPOSITION = "dischargeDisposition"
+final String DEPARTMENT = "department:"
+
+
+final Map PROFILE_TYPES = [
+    (TREATMENT_SERVICE)    : LaborFindingLaborValue.MULTI_VALUE_REFERENCES,
+    (TYPE)                 : LaborFindingLaborValue.CATALOG_ENTRY_VALUE,
+    (DISCHARGE_DISPOSITION): LaborFindingLaborValue.CATALOG_ENTRY_VALUE,
+    (DEPARTMENT)           : LaborFindingLaborValue.MULTI_VALUE_REFERENCES
+]
+
 encounter {
 
   if (isFakeEpisode(context.source)) {
@@ -32,6 +58,12 @@ encounter {
   }
 
   id = "Encounter/" + context.source[episode().id()]
+
+  final def mapping = context.source[medication().laborMappings()].find { final def lm ->
+    lm[LaborMapping.LABOR_FINDING][LaborFinding.LABOR_METHOD][CODE] == "Encounter_profile"
+  }
+
+  final Map<String, Object> lflvMap = getLflvMap(mapping, PROFILE_TYPES)
 
   meta {
     profile "http://hl7.org/fhir/us/core/StructureDefinition/us-core-encounter"
@@ -52,11 +84,10 @@ encounter {
     }
   }
 
-  status = Encounter.EncounterStatus.UNKNOWN
+  status = mapStatus(context.source[episode().status()] as String)
 
-  if (context.source[episode().stayType().code()]) {
+  if (context.source[episode().stayType()]) {
     class_ {
-      system = FhirUrls.System.Episode.StayType.BASE_URL
       code = context.source[episode().stayType().code()]
       display = context.source[episode().stayType().multilinguals()].find { final def ml ->
         ml[Multilingual.SHORT_NAME] != null && ml[Multilingual.LANGUAGE] == "en"
@@ -119,6 +150,47 @@ encounter {
       }
     }
   }
+
+  if (lflvMap.containsKey(DEPARTMENT)) {
+    lflvMap.get(DEPARTMENT).each { final def valueRef ->
+      if (valueRef && valueRef[ValueReference.ORGANIZATION_VALUE]) {
+        extension {
+          url = "https://fhir.iqvia.com/patientfinder/extension/department-organization"
+          valueReference {
+            reference = "Organization/" + valueRef[ValueReference.ORGANIZATION_VALUE][OrganisationUnit.ID]
+          }
+        }
+      }
+    }
+  }
+
+  if (lflvMap.containsKey(TYPE)) {
+    type {
+      lflvMap.get(TYPE).each { final def entry ->
+        coding {
+          code = entry[CODE] as String
+          display = entry[MULTILINGUALS].find { final def ml ->
+            ml[Multilingual.SHORT_NAME] != null && ml[Multilingual.LANGUAGE] == "en"
+          }?.getAt(Multilingual.SHORT_NAME)
+        }
+      }
+    }
+  }
+
+  if (lflvMap.containsKey(DISCHARGE_DISPOSITION)) {
+    hospitalization {
+      dischargeDisposition {
+        lflvMap.get(DISCHARGE_DISPOSITION).each { final def entry ->
+          coding {
+            code = entry[CODE] as String
+            display = entry[MULTILINGUALS].find { final def ml ->
+              ml[Multilingual.SHORT_NAME] != null && ml[Multilingual.LANGUAGE] == "en"
+            }?.getAt(Multilingual.SHORT_NAME)
+          }
+        }
+      }
+    }
+  }
 }
 
 static boolean isFakeEpisode(final def episode) {
@@ -132,4 +204,42 @@ static boolean isFakeEpisode(final def episode) {
 
   final def fakeId = episode[Episode.ID_CONTAINER]?.find { (it[PSN] as String).toUpperCase().startsWith("FAKE") }
   return fakeId != null
+}
+
+static Map<String, Object> getLflvMap(final def mapping, final Map<String, String> types) {
+  final Map<String, Object> lflvMap = [:]
+  if (!mapping) {
+    return lflvMap
+  }
+
+  types.each { final String lvCode, final String lvType ->
+    final def lflvForLv = mapping[LaborMapping.LABOR_FINDING][LaborFinding.LABOR_FINDING_LABOR_VALUES].find { final def lflv ->
+      lflv[LaborFindingLaborValue.CRF_TEMPLATE_FIELD][CrfTemplateField.LABOR_VALUE][CODE] == lvCode
+    }
+
+    if (lflvForLv && lflvForLv[lvType]) {
+      lflvMap[(lvCode)] = lflvForLv[lvType]
+    }
+  }
+  return lflvMap
+}
+
+static Encounter.EncounterStatus mapStatus(final String cxxStatus) {
+  switch (cxxStatus) {
+    case EpisodeStatus.FINISHED.toString():
+      return Encounter.EncounterStatus.FINISHED
+    case EpisodeStatus.ARRIVED.toString():
+      return Encounter.EncounterStatus.ARRIVED
+    case EpisodeStatus.CANCELED.toString():
+      return Encounter.EncounterStatus.CANCELLED
+    case EpisodeStatus.TRIAGED.toString():
+      return Encounter.EncounterStatus.TRIAGED
+    case EpisodeStatus.PLANNED.toString():
+      return Encounter.EncounterStatus.PLANNED
+    case EpisodeStatus.IN_PROGRESS.toString():
+      return Encounter.EncounterStatus.INPROGRESS
+
+    default:
+      return Encounter.EncounterStatus.UNKNOWN
+  }
 }